Payment Model

+++ migrations/2024-12-18-130213_create_all/up.sql
@@ -30,7 +30,7 @@ CREATE TABLE "payments"(
 	"name" TEXT NOT NULL,
 	"description" TEXT NOT NULL,
 	"status" TEXT NOT NULL,
-	"balance" NUMERIC NOT NULL,
+	"balance" FLOAT NOT NULL,
 	"created_date" TIMESTAMP NOT NULL,
 	"updated_date" TIMESTAMP NOT NULL
 );
